+++ src/Database/migrations/2021_10_09_100011_create_containers_table.php
@@ -15,7 +15,7 @@ class CreateContainersTable extends Migration
     {
         Schema::create('containers', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('structure_id')->index();
+            $table->unsignedBigInteger('structure_id')->index();
             $table->string('cvar_1')->nullable();
             $table->string('cvar_2')->nullable();
             $table->string('cvar_3')->nullable();
