+++ src/Database/migrations/2021_10_09_100031_create_property_translatations_table.php
@@ -15,7 +15,7 @@ class CreatePropertyTranslationsTable extends Migration
     {
         Schema::create('property_translations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('property_id')->index();
+            $table->unsignedBigInteger('property_id')->index();
             $table->unsignedInteger('language_id')->index();
             $table->boolean('status')->default(1);
             $table->string('name');
