+++ src/Database/migrations/2021_10_09_100042_create_page_translatations_table.php
@@ -15,7 +15,7 @@ class CreatePageTranslationsTable extends Migration
     {
         Schema::create('page_translations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('page_id')->index();
+            $table->unsignedBigInteger('page_id')->index();
             $table->unsignedInteger('language_id')->index();
             $table->boolean('status')->default(1);
             $table->string('name');
